fix: progressbar hide too early
