+++ package.json
@@ -32,12 +32,12 @@
     "contentful:import": "contentful space import --config contentful.json"
   },
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.4",
+    "@apollo/react-hooks": "^3.1.5",
     "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
-    "@mdx-js/mdx": "^1.5.8",
-    "@mdx-js/react": "^1.5.8",
-    "@now/node": "^1.5.0",
+    "@mdx-js/mdx": "^1.5.9",
+    "@mdx-js/react": "^1.5.9",
+    "@now/node": "^1.5.1",
     "@theme-ui/color": "^0.3.1",
     "@theme-ui/components": "^0.3.1",
     "@theme-ui/preset-deep": "^0.3.0",
@@ -49,34 +49,34 @@
     "contentful-cli": "^0.36.2",
     "cross-env": "^7.0.2",
     "framer-motion": "^1.10.3",
-    "gatsby": "^2.20.17",
-    "gatsby-image": "^2.3.2",
+    "gatsby": "^2.20.29",
+    "gatsby-image": "^2.3.4",
     "gatsby-paginate": "^1.1.1",
-    "gatsby-plugin-catch-links": "^2.2.1",
-    "gatsby-plugin-emotion": "^4.2.1",
+    "gatsby-plugin-catch-links": "^2.2.3",
+    "gatsby-plugin-emotion": "^4.2.2",
     "gatsby-plugin-eslint": "^2.0.8",
-    "gatsby-plugin-feed": "^2.4.1",
-    "gatsby-plugin-google-analytics": "^2.2.2",
+    "gatsby-plugin-feed": "^2.4.3",
+    "gatsby-plugin-google-analytics": "^2.2.4",
     "gatsby-plugin-mailchimp": "^5.1.2",
-    "gatsby-plugin-manifest": "^2.3.3",
-    "gatsby-plugin-mdx": "^1.1.7",
-    "gatsby-plugin-nprogress": "^2.2.1",
-    "gatsby-plugin-react-helmet": "^3.2.2",
-    "gatsby-plugin-sharp": "^2.5.3",
+    "gatsby-plugin-manifest": "^2.3.6",
+    "gatsby-plugin-mdx": "^1.1.10",
+    "gatsby-plugin-nprogress": "^2.2.3",
+    "gatsby-plugin-react-helmet": "^3.2.4",
+    "gatsby-plugin-sharp": "^2.5.6",
     "gatsby-plugin-theme-ui": "^0.3.0",
-    "gatsby-plugin-typescript": "^2.3.1",
+    "gatsby-plugin-typescript": "^2.3.3",
     "gatsby-plugin-webpack-bundle-analyzer": "^1.0.5",
     "gatsby-remark-check-links": "^2.1.0",
-    "gatsby-remark-copy-linked-files": "^2.2.1",
+    "gatsby-remark-copy-linked-files": "^2.2.3",
     "gatsby-remark-external-links": "0.0.4",
-    "gatsby-remark-images": "^3.2.2",
+    "gatsby-remark-images": "^3.2.5",
     "gatsby-remark-numbered-footnotes": "^1.0.1",
-    "gatsby-remark-responsive-iframe": "^2.3.1",
-    "gatsby-remark-smartypants": "^2.2.1",
+    "gatsby-remark-responsive-iframe": "^2.3.3",
+    "gatsby-remark-smartypants": "^2.2.3",
     "gatsby-remark-typescript": "^0.2.2",
-    "gatsby-source-filesystem": "^2.2.2",
-    "gatsby-transformer-sharp": "^2.4.3",
-    "gatsby-transformer-yaml": "^2.3.1",
+    "gatsby-source-filesystem": "^2.2.4",
+    "gatsby-transformer-sharp": "^2.4.6",
+    "gatsby-transformer-yaml": "^2.3.3",
     "gh-pages": "^2.2.0",
     "graphql": "^14.6.0",
     "instant.page": "^3.0.0",
@@ -84,17 +84,17 @@
     "moment-timezone": "^0.5.28",
     "node-fetch": "^2.6.0",
     "prism-react-renderer": "^1.0.2",
-    "query-string": "^6.12.0",
+    "query-string": "^6.12.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-helmet": "^5.2.1",
     "react-live": "^2.2.2",
     "react-medium-image-zoom": "^3.1.2",
     "react-outside-click-handler": "1.0.0",
-    "react-query": "^1.2.1",
+    "react-query": "^1.2.9",
     "react-transition-group": "^4.3.0",
     "remark-slug": "^5.1.2",
-    "semver": "^7.2.2",
+    "semver": "^7.3.2",
     "theme-ui": "^0.3.1",
     "tiny-glob": "^0.2.6",
     "typeface-merriweather": "0.0.72",
@@ -103,11 +103,11 @@
   "devDependencies": {
     "@commitlint/cli": "^8.3.5",
     "@commitlint/config-conventional": "^8.3.4",
-    "@graphql-codegen/cli": "^1.13.2",
-    "@graphql-codegen/introspection": "^1.13.2",
-    "@graphql-codegen/typescript": "^1.13.2",
-    "@graphql-codegen/typescript-operations": "^1.13.2",
-    "@types/lodash": "^4.14.149",
+    "@graphql-codegen/cli": "^1.13.3",
+    "@graphql-codegen/introspection": "^1.13.3",
+    "@graphql-codegen/typescript": "^1.13.3",
+    "@graphql-codegen/typescript-operations": "^1.13.3",
+    "@types/lodash": "^4.14.150",
     "@types/node": "^12.12.34",
     "@types/react": "^16.9.34",
     "@types/react-dom": "^16.9.6",
@@ -118,7 +118,7 @@
     "eslint": "^6.8.0",
     "eslint-config-prettify-ts-react": "^0.1.8",
     "husky": "^4.2.5",
-    "lint-staged": "^10.1.3",
+    "lint-staged": "^10.1.7",
     "prettier": "~1.19.1",
     "typescript": "~3.8.3"
   },
